requirements.txt
openai==0.27.8
requests==2.31.0
fastapi==0.100.1
uvicorn==0.23.1
python-dotenv==1.0.0
PyPDF2==3.0.1
